battle card/requirements.txt
Flask==3.0.3
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
google-generativeai==0.5.4  # Use the actual version if different
pandas==2.0.2
openpyxl==3.1.5
markupsafe==2.1.3
